refactor document and connector into MongoAdapter class
