+++ package.json
@@ -2,7 +2,7 @@
   "name": "mongo-adapter",
   "version": "1.0.0",
   "description": "Cloudboost MongoDB adapter",
-  "main": "dist/adapter.js",
+  "main": "dist/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "babel src --out-dir dist"
