--- src/adapter.js
@@ -1,9 +1,0 @@
-const connector = require('./connector');
-const document = require('./document');
-const util = require('./util');
-
-module.exports = {
-  connector,
-  document,
-  util
-};
