+++ src/index.js
@@ -1,0 +1,11 @@
+import MongoAdapter from './MongoAdapter';
+import Document from './Document';
+import util from './util';
+import errors from './errors';
+
+module.exports = {
+  MongoAdapter,
+  Document,
+  util,
+  errors
+};
